refactor: page 파일 이름 소문자 변경
